CI: update toolchain to v1.59.0
In the commit `b9501de55fe8532dfe5ea7fccfcf9a5241a17db9', we see that
the minimum used version of the toolchain is considered to be v1.59.0.

Also, there is no need to strip the final executable, as this is handled
by Cargo anymore, while building release targets.  But, it doesn't hurt
to leave the relevant command in the `Makefile' either.